backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
dashscope>=1.20.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# RAG系统新增依赖
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
GitPython==3.1.40
schedule==1.2.0 